+++ src/tilda/types/Type_JsonPrimitive.java
@@ -20,7 +20,7 @@ import tilda.enums.ColumnType;
 
 public class Type_JsonPrimitive extends ColumnDefinition
   {
-    public Type_JsonPrimitive(String SchemaName, String TableName, String ColumnName, long Count, String Description)
+    public Type_JsonPrimitive(String SchemaName, String TableName, String ColumnName, int Count, String Description)
       {
         super(SchemaName, TableName, ColumnName, Count, ColumnType.JSON, false, Description);
       }
