+++ src/tilda/types/Type_JsonPrimitiveNull.java
@@ -19,7 +19,7 @@ package tilda.types;
 
 public class Type_JsonPrimitiveNull extends Type_JsonPrimitive implements Nullable
   {
-    public Type_JsonPrimitiveNull(String SchemaName, String TableName, String ColumnName, long Count, String Description)
+    public Type_JsonPrimitiveNull(String SchemaName, String TableName, String ColumnName, int Count, String Description)
       {
         super(SchemaName, TableName, ColumnName, Count, Description);
       }
